pyproject.toml
[tool.poetry]
name = "up-python"
version = "1.5.8-dev"
description = "Language specific uProtocol library for building and using UUri, UUID, UAttributes, UTransport, and more."
authors = ["Neelam Kushwah <neelam.kushwah@gm.com>"]
license = "The Apache License, Version 2.0"
readme = "README.adoc"
repository = "https://github.com/eclipse-uprotocol/up-python"
packages = [{ include = "uprotocol" },
    { include = "tests" },
    { include = "scripts" }

]

[tool.poetry.dependencies]
python = "^3.8"
cloudevents = "*"
multimethod = "*"
gitpython = ">=3.1.41"
googleapis-common-protos = ">=1.56.4"
protobuf = "4.24.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
coverage = "^5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"